v1.0: ready for release
