+++ README.md
@@ -1,8 +1,68 @@
-# React + Vite
+<h1 align="center">Draw-it-out: Online Whiteboard App</h1> 
 
-This template provides a minimal setup to get React working in Vite with HMR and some ESLint rules.
+<div align="center"><img src=".src/assets/images/webLogo.png" width="200px" height="200px"/></div>
 
-Currently, two official plugins are available:
+# Introduction
+Draw-it-out is an online whiteboard tool built using ReactJS and TailwindCSS. It allows users to draw, sketch, and take snapshots of their creations. The intuitive interface and responsive design make it easy to use across various devices.
 
-- [@vitejs/plugin-react](https://github.com/vitejs/vite-plugin-react/blob/main/packages/plugin-react/README.md) uses [Babel](https://babeljs.io/) for Fast Refresh
-- [@vitejs/plugin-react-swc](https://github.com/vitejs/vite-plugin-react-swc) uses [SWC](https://swc.rs/) for Fast Refresh
+# How to Contribute
+We welcome contributions from the open-source community to improve ScanVerse. Follow the steps below to contribute:
+
+1. Fork the repository on GitHub.
+
+1. Clone the forked repository to your local machine.
+
+`git clone https://github.com/your-username/Draw-it-out.git` </br>
+
+3. Create a new branch and switch to it.
+
+`git checkout -b <branch_name>` </br>
+
+4. Make the necessary changes or add new features.
+
+5. Commit your changes with a descriptive message.
+
+`git commit -m "commit_message"` </br>
+
+6. Push your changes to your forked repository.
+
+`git push origin feature/<my-contribution>` </br>
+
+7. Open a Pull Request (PR) from your forked repository to the main repository.
+
+8. Provide a clear and detailed description of your changes in the PR.
+
+9. Wait for the code reviewers to review your PR and address any feedback if necessary.
+
+10. Once approved, your changes will be merged into the main repository.
+
+# Tech Stack
+
+* **ReactJS**
+* **TailwindCSS** 
+
+# Author
+
+* **_Shubham Singodiya_** - Lead developer and creator of the whiteboard web application.
+
+We appreciate our team member's hard work and dedication to make ScanVerse a reality.
+
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+
+# Releases
+
+* **Release 1.0** : March 15, 2024 @16:32 [First phase of application completed.]
+  - Features: Added snapshot functionality and fixed minor bugs.
+
+* **Release 0.0** : March 14, 2024 @22:31 [Zeroth phase of application completed.]
+  - Features: Drawing functionality with custom thickness of the brush and custom color.
+
+<!--- Acknowledgements: We would like to thank all the contributors who have helped in the development of ScanVerse. I would greatly appreciate your support and contributions -->
+
+
+# Message from PA
+Thank you for visiting Draw-it-out!💝
+</br>Feel free to explore the code, contribute, and provide feedback.
+
+
+
