+++ src/components/Menu.jsx
@@ -1,0 +1,60 @@
+import { PiPencilSimpleFill } from "react-icons/pi";
+import { FaFeatherPointed } from "react-icons/fa6";
+import { RiScreenshot2Fill } from "react-icons/ri";
+import { useState } from "react";
+import { takeSnapshot } from "../utils/takeSnapshot.js";
+
+// eslint-disable-next-line react/prop-types
+const Menu = ({ drawable, setDrawable, thickness, setThickness, setColor, canvasRef }) => {
+  const [pencilWidth, setPencilWidth] = useState(false);
+
+  return (
+    <>
+      <div className="max-w-[90%] flex-wrap	 tools bg-[#CBCCCF] shadow-mdm flex justify-center items-stretch gap-[1rem] md:gap-[2rem] px-[2rem] py-4 rounded-[0.6rem]">
+        <PiPencilSimpleFill
+          className={`text-[2rem] md:text-[3rem] p-[0.5rem] md:p-[0.8rem] shadow-vsm rounded-[0.5rem] cursor-pointer hover:bg-[#B7BABF] ${
+            drawable ? "bg-gray-400" : ""
+          }`}
+          onClick={() => setDrawable(!drawable)}
+          title="Draw"
+        />
+        <FaFeatherPointed
+          className={`text-[2rem] md:text-[3rem] p-[0.5rem] md:p-[0.8rem] shadow-vsm rounded-[0.5rem] cursor-pointer hover:bg-[#B7BABF] ${
+            pencilWidth ? "bg-gray-400" : ""
+          }`}
+          onClick={() => setPencilWidth(!pencilWidth)}
+          title="Brush Thickness"
+        />
+        {pencilWidth && (
+          <input
+            type="range"
+            name="thickness"
+            id="thickness"
+            value={thickness || 10}
+            min={1}
+            max={100}
+            onChange={(e) => setThickness(e.target.value)}
+            className="cursor-pointer"
+          />
+        )}
+        <div className="p-[1rem] px-[1.5rem] rounded-[0.5rem] relative shadow-vsm hover:bg-[#B7BABF] cursor-pointer">
+          <input
+            type="color"
+            name="color"
+            id="color"
+            title="Color Picker"
+            onChange={(e) => setColor(e.target.value)}
+            className={`bg-[#CBCCCF] p-[0.5rem] shadow-vsm rounded-[0.5rem] cursor-pointer outline-none hover:bg-[#B7BABF] flex-[0.5] w-full h-full z-[5] absolute top-0 left-0`}
+          />
+        </div>
+        <RiScreenshot2Fill
+          className={`text-[2rem] md:text-[3rem] p-[0.5rem] md:p-[0.8rem] shadow-vsm rounded-[0.5rem] cursor-pointer hover:bg-[#B7BABF]`}
+          onClick={()=>takeSnapshot(canvasRef)}
+          title="Snapshot"
+        />
+      </div>
+    </>
+  );
+};
+
+export default Menu;
